refactor(media-library): extract deletion flow into a hook
Consolidates the scattered deletion logic — pending-deletion/dialog state, the
Delete-hotkey and focus-scope effects, timeline-impact computations, and the
confirm/cancel handlers — into use-media-library-deletion.ts (also home to the
shared PendingLibraryDeletion type). Verbatim behaviour; stable refs added to
effect deps.

media-library.tsx: 1546 -> 1431 lines (cumulative 1882 -> 1431, -24%).
